requirements.txt
fire==0.3.1
pygithub==1.51